# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/category": {
            "put": {
                "description": "피자정보를 update하는 api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update a pizza in category",
                "parameters": [
                    {
                        "description": "PizzaCategory Info",
                        "name": "structs.RequestPizzaCategoryBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.RequestPizzaCategoryBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PizzaCategory"
                        }
                    }
                }
            },
            "post": {
                "description": "피자종류를 추가하는 api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "add a pizza in category",
                "parameters": [
                    {
                        "description": "PizzaCategory Info",
                        "name": "structs.RequestPizzaCategoryBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.RequestPizzaCategoryBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PizzaCategory"
                        }
                    }
                }
            },
            "delete": {
                "description": "피자정보를 update하는 api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "delete a pizza in category",
                "parameters": [
                    {
                        "description": "delete",
                        "name": "DeleteRequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admincontroller.DeleteRequestBody"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/admin/order/update": {
            "post": {
                "description": "피자주문정보에서 주문접수, 조리, 배달완료 등 상태를 update하는 API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update a status in order",
                "parameters": [
                    {
                        "description": "update a status in order",
                        "name": "OrderStatusBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admincontroller.OrderStatusBody"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/health": {
            "get": {
                "description": "health check",
                "summary": "health check",
                "responses": {}
            }
        },
        "/pizza/order": {
            "post": {
                "description": "피자를 주문하는 API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "order a pizza",
                "parameters": [
                    {
                        "description": "RequestPizzaOrderBody Info",
                        "name": "structs.RequestPizzaOrderBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.RequestPizzaOrderBody"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/pizza/order/id/{id}": {
            "get": {
                "description": "주문번호를 통해서 주문정보를 받아볼 수 있는 API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "주문번호를 통해서 주문정보를 받아볼 수 있는 API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/pizza/order/{name}/{phone}": {
            "get": {
                "description": "이름과 전화번호를 통해서 주문내역을 확인할 수 있는 API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "이름과 전화번호를 통해서 주문내역을 확인할 수 있는 API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user phone",
                        "name": "phone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "admincontroller.DeleteRequestBody": {
            "type": "object",
            "required": [
                "pizzaId"
            ],
            "properties": {
                "pizzaId": {
                    "type": "string"
                }
            }
        },
        "admincontroller.OrderStatusBody": {
            "type": "object",
            "required": [
                "id",
                "status"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.PizzaCategory": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "deleted_At": {
                    "type": "string"
                },
                "des": {
                    "type": "string"
                },
                "l": {
                    "type": "integer"
                },
                "limit_Order": {
                    "type": "integer"
                },
                "m": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "order_status": {
                    "type": "boolean"
                },
                "updated_At": {
                    "type": "string"
                }
            }
        },
        "structs.RequestOrderInfo": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "personId": {
                    "type": "string"
                },
                "pizzaId": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "structs.RequestOrderPersonInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "pizzaId": {
                    "type": "string"
                }
            }
        },
        "structs.RequestPizzaCategoryBody": {
            "type": "object",
            "required": [
                "des",
                "l",
                "limit_Order",
                "m",
                "name",
                "order_status"
            ],
            "properties": {
                "des": {
                    "type": "string"
                },
                "l": {
                    "type": "integer"
                },
                "limit_Order": {
                    "type": "integer"
                },
                "m": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "order_status": {
                    "type": "boolean"
                }
            }
        },
        "structs.RequestPizzaOrderBody": {
            "type": "object",
            "properties": {
                "orderInfo": {
                    "$ref": "#/definitions/structs.RequestOrderInfo"
                },
                "personInfo": {
                    "$ref": "#/definitions/structs.RequestOrderPersonInfo"
                }
            }
        }
    }
}